feat(channel): mark articles not as spam when setting topic channels
and split tests into multipe files
